requirements.txt
crewai==0.11.0
python-dotenv==1.0.0
python-docx==1.0.1
duckduckgo-search==3.9.3
langchain==0.0.329
langchain-openai==0.0.2.post1
openai>=1.6.1 